+++ MinimOsd_Extra/Config.h
@@ -9,19 +9,21 @@
 
 #define VERSION "2.4" 
 
-//#define DEBUG
+// **************************
+#define DEBUG
 
 
 #define STARTUP_SCREEN 0
-#define AUTOBAUD 1
-//#define SHOW_FONT 1 not work
-#define USE_SENSORS 1
+
+#define MAVLINK_CONFIG 1
+#define MAVLINK_READ_EEPROM 1
+
 
 #ifndef DEBUG // don't fit
 //#define FONT_UPLOAD 1
 #define USE_SETUP 1
-#define MAVLINK_CONFIG 1
-#define MAVLINK_READ_EEPROM 1
+#define AUTOBAUD 1
+#define USE_SENSORS 1
 #endif
 
 //#define PWM_IN_INTERRUPT 1