+++ MinimOsd_Extra/misc.h
@@ -3,7 +3,7 @@
 #ifdef MAVLINK_CONFIG
 
 
-void mavlink_return_packet(uint8_t id, uint8_t len);
+void mavlink_return_packet(uint8_t id, uint8_t len, uint8_t crc);
 
 struct Mav_conf { // needs to fit in 253 bytes
     byte magick[4]; // = 0xee 'O' 'S' 'D'
@@ -38,7 +38,7 @@ void parse_osd_packet(byte *p){
         case 'r':
             eeprom_read_len((byte *)&c->data, (uint16_t)(c->id) * 128,  c->len );
 
-            mavlink_return_packet(MAVLINK_MSG_ID_ENCAPSULATED_DATA, MAVLINK_MSG_ID_ENCAPSULATED_DATA_LEN); // send packet back
+            mavlink_return_packet(MAVLINK_MSG_ID_ENCAPSULATED_DATA, MAVLINK_MSG_ID_ENCAPSULATED_DATA_LEN, MAVLINK_MSG_ID_ENCAPSULATED_DATA_CRC); // send packet back
             break;
             
 #endif
@@ -61,18 +61,23 @@ void parse_osd_packet(byte *p){
     uint8_t msgid;   ///< ID of message in payload
     uint64_t payload64[(MAVLINK_MAX_PAYLOAD_LEN+MAVLINK_NUM_CHECKSUM_BYTES+7)/8];
 */
-void mavlink_return_packet(uint8_t id, uint8_t len) {
+void mavlink_return_packet(uint8_t id, uint8_t len, uint8_t crc) {
     uint16_t checksum;
     
     msg.m.magic  = MAVLINK_STX;
     msg.m.len    = len;
     msg.m.msgid  = id;
+    msg.m.sysid  = mavlink_system.sysid;
+    msg.m.compid = MAV_COMP_ID_CAMERA; // stole this id
 
-    checksum = crc_calculate(((const uint8_t*)&msg.m)+1, MAVLINK_CORE_HEADER_LEN + MAVLINK_MSG_ID_ENCAPSULATED_DATA_LEN);
+    checksum = crc_calculate(((const uint8_t*)&msg.m.len), MAVLINK_CORE_HEADER_LEN + len);
 #if MAVLINK_CRC_EXTRA
-    crc_accumulate(MAVLINK_MSG_ID_ENCAPSULATED_DATA_CRC, &checksum);
+    crc_accumulate(crc, &checksum);
 #endif
-    _mavlink_send_uart((mavlink_channel_t)0, (const char *)&msg.m, (uint16_t)(MAVLINK_NUM_HEADER_BYTES + MAVLINK_MSG_ID_ENCAPSULATED_DATA_LEN) );
+//    serial_hex_dump((byte *)&msg.m.magic, MAVLINK_NUM_HEADER_BYTES + len);
+
+    _mavlink_send_uart((mavlink_channel_t)0, (const char *)&msg.m.magic, MAVLINK_NUM_HEADER_BYTES);
+    _mavlink_send_uart((mavlink_channel_t)0, (const char *)&msg.m.payload64, len );
     _mavlink_send_uart((mavlink_channel_t)0, (const char *)&checksum, 2);    
 }