+++ libraries/GCS_MAVLink/GCS_MAVLink.cpp
@@ -21,7 +21,7 @@ version 2.1 of the License, or (at your option) any later version.
 extern BetterStream	*mavlink_comm_0_port;
 //BetterStream	*mavlink_comm_1_port;
 
-extern mavlink_system_t mavlink_system = {12,1}; //modified
+extern mavlink_system_t mavlink_system; //modified
 
 uint8_t mavlink_check_target(uint8_t sysid, uint8_t compid)
 {